[feature] 홈탭 절약기록 리스트 등록, 수정, 삭제 기능 구현

+++ domain/src/main/java/com/yapp/android2/domain/entity/LoginResponse.kt
@@ -7,6 +7,6 @@ data class LoginResponse(override val data: Data) : BaseResponse<LoginResponse.D
         val accessToken: String?,
         val nickName: String?,
         val refreshToken: String?,
-        val userId: Int?
+        val userId: Long?
     )
 }
